refactor: extract shared util function from media handler
